% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_condition.R
\name{check_condition}
\alias{check_condition}
\title{rust like unwrapping of result. Useful to keep error handling on the R side.}
\usage{
check_condition(value_or_condition, call = sys.call(1L), ...)
}
\arguments{
\item{call}{a context call for what triggered the error, passed to abort(footer=). abort(call=) will annoyingly not display anonymous funcitons.}

\item{...}{any other param to rlang::abort}

\item{result}{a list with two elements named 'ok' and 'err'. One and only one of them must NULL or missing.}
}
\value{
eiter the ok value or no return but throw error by the err string
}
\description{
rust like unwrapping of result. Useful to keep error handling on the R side.
}
\examples{
unwrap(rust_helloresult(TRUE))
}
