% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_unwrap.R
\name{unwrap}
\alias{unwrap}
\title{rust like unwrapping of result. Useful to keep error handling on the R side.}
\usage{
unwrap(result, call = sys.call(1L), ...)
}
\arguments{
\item{result}{a list with two elements named 'ok' and 'err'. One and only one of them must NULL or missing.}

\item{call}{a context call for what triggered the error, passed to abort(footer=). abort(call=) will annoyingly not display anonymous funcitons.}

\item{...}{any other param to rlang::abort}
}
\value{
eiter the ok value or no return but throw error by the err string
}
\description{
rust like unwrapping of result. Useful to keep error handling on the R side.
}
\examples{
unwrap(extendr_result:::rust_hello_result(TRUE))
}
